Refactor project frontend API and fix backend project API
This commit refactors the project frontend API by adding tests and fixes an issue with the backend project API where the user ID should be a number.

Note: Please remove any meta information like issue references, tags, or author names from the commit message.
