+++ frontend/src/utils/client.ts
@@ -5,7 +5,6 @@ import {
   ApolloLink,
   concat,
 } from '@apollo/client';
-import { useMemo } from 'react';
 
 const httpLink = new HttpLink({
   uri: process.env.NEXT_PUBLIC_GRAPHQL_URL,
@@ -24,13 +23,19 @@ const authMiddleware = new ApolloLink((operation, forward) => {
   }
   return forward(operation);
 });
+const client = new ApolloClient({
+  link: concat(authMiddleware, httpLink),
+  cache: new InMemoryCache(),
+});
 
-export const useClient = () => {
-  const client = useMemo(() => {
-    return new ApolloClient({
-      link: concat(authMiddleware, httpLink),
-      cache: new InMemoryCache(),
-    });
-  }, []);
-  return client;
-};
+export default client;
+
+// export const useClient = () => {
+//   const client = useMemo(() => {
+//     return new ApolloClient({
+//       link: concat(authMiddleware, httpLink),
+//       cache: new InMemoryCache(),
+//     });
+//   }, []);
+//   return client;
+// };
